requirements.txt
pytest==3.6.1
pytest-allure-adaptor==1.7.10
PyHamcrest==1.9.0
selenium==3.12.0
pytest-xdist==1.23.
slackclient==1.3.0
Pillow==5.3.0